+++ Common/Panel.cpp
@@ -17,7 +17,7 @@ void Panel::Add(Control* control) {
 	controls.push_back(control);
 }
 
-void Panel::draw(Graphics& g, int x, int y, size_t z) {
+void Panel::draw(Graphics& g, short x, short y, size_t z) {
 	Control::draw(g, x, y, z);
 	int i;
 	for (i = 0; i < controls.size() - 1; i++) {
