+++ Common/RadioBox.cpp
@@ -11,6 +11,6 @@ RadioBox::~RadioBox()
 {
 }
 
-void RadioBox::draw(Graphics& g, int x, int y, size_t z) {
+void RadioBox::draw(Graphics& g, short x, short y, size_t z) {
 
 }
